Sort Decode As dissector list case-insensitively
When selecting a dissector in the Decode As dialog box, the combo box of
dissectors is sorted by the ASCII of the dissector description (or protocol
name). That's not very user-friendly because it's case-sensitive; protocols
starting with a lowercase letter appear at the end of the list.

Put the dissectors in a QList and sort the list using localeAwareCompare,
instead of using a QMap and relying on its implicit sorting by key. This
sorts them case-insensitively, and makes it easier for the user to find
protocols in the list.