TLS: Handle cleartext handshakes after CCS
Everything in the same TLS session is supposed to be encrypted after
the Change Cipher Spec, including a subsequent new handshake for
TLS renegotiation. Some clients have a buggy implementation of
renegotiation and send the new handshake in the clear.

Change around the encrypted handshake heuristics some. If we're
after the CCS, but it looks like a new ClientHello or ServerHello,
assume this is a new cleartext TLS handshake and reset the CCS state.

Fix #18867