Support Wi-SUN EAPOL Key Data dissection
The Wi-SUN FAN specification describes the format of the EAPOL-Key frame
in section 6.5.2.2 (Authentication and PMK Installation Flow):

  Descriptor Type = 2
  Key Information:
    1. Key Descriptor Version = 2
    2. Key Type = 0
    3. Install = 0
    4. Key Ack = 0
    5. Key MIC = 0
    6. Secure = 0
    7. Error = 0
    8. Request = 1
    9. Encrypted Key Data = 0
    10. SMK Message = 0
    11. Reserved = 0
  Key Length = 0
  Key Replay Counter = see [IEEE802.11] section 11.6.2.
  Key Nonce = 0
  EAPOL-Key IV = 0
  Key RSC = 0
  Key MIC = 0
  Key Data Length = length of Key Data field in octets.
  Key Data = PMKID KDE if the PMK is live, PTKID KDE if the PTK is live, GTKL
  KDE, Node Role KDE, and LGTKL KDE.

The current dissector will try do decrypt if the Key Type is 0 while the
Encrypted Key Data is unset, which appears to be for supporting
non-standard WPA implementations. The Key Data is not encrypted in
Wi-SUN, so a workaround is made to dissect the Key Data if the Key
Length is 0.