Add new dissector for GSM RLP protocol (3GPP TS 24.022)
This protocol is used in the user plane of non-transparent CSD (Circuit
Switched Data) calls in GSM networks. RLP frames are sent over the Um
air interface, and are sent as modified V.110 frames over 64k TDM
channels in the back-haul/core network.  For modern implementations,
this means in RFC4040  RTP CLEARMODE.

As there's no V.110 decoder in wireshark, we cannot connect the RLP
decoder to that.  However, we hook it up to the GSMTAP dissector to
enable other software to pass the decoded RLP frames into wireshark.