String search may fail when partial matches occur
When searching for string matches in packet list and details, save index of
next possible start location inside a partial match and rewind to that
position if the ongoing match is not successful.

While here, also terminate search when the search string is longer than
the remainder of the text string being matched against.

As suggested by John Thacker, use strstr() for case sensitive search.

Full bug description:

Currently, searching for strings in "Packet list" and "Packet detail" may fail
to find matches if a partial match is encountered. Examples of both are present
in nfsv4.1_pnfs.cap in the Wireshark sample trace collection.

Searching for "Win=29200 Len=0" in Packet list finds frame 1, based on:

    880 → 2049 [SYN] Seq=0 Win=29200 Len=0 MSS=1460 SACK_PERM TSval=360391592 TSecr=0 WS=128

However, searching for jsut "0 Len=0" skips this frame when searching.

Similarly, searching for "netapp-26" in Packet detail will find many frames
(if TCP port 2049 is decoded as RPC) based on the RPC credentials:

    Machine Name: netapp-26

Similarly, if searching for "p-26", no frames are found.

The problem is how match_summary_line() and match_subtree_text() will start a
comparison against the search string based on the first character and scan
ahead to check if each subsequent character also matches. However, if there is
no match, the search continues only after the partial match.

In practice, if "p-26" is against "Machine Name: netapp-26" will:

- Test against against characters, but find no match until the first 'p'.
- Find that the first 'p' matches the leading 'p' in the search string.
- Compare the second 'p' against '-' in the search string, which fails.
- Continue checking the second 'p' where neither '-', '2' nor '6' match the
  leading 'p' in the search string.

The proposed fix will, when a partial match is in progress, store the first
location where a new match might occur. If the partial match is not
successful, the search is restarted at that stored position.

As far as I can tell, other match_xxx functions in file.c do not have
share this problem.