Add handling for can error frames to blf file format
There are three different can error objects in the blf format corresponding
to different generations of the vector tools using the blf file format.
All three are parsed and the errors are mapped to socketcan error frames
which can already be displayed. For errors with valid can core error
information this is mapped to the error class of socketcan.