snort: don't look at the encap type when writing.
When writing a file format that doesn't support multiple encapsulations
in a file, libwiretap returns an error if we try to do that.  We don't
need to reach into the wiretap dumper structure to do that check
ourselves.

(Not that we were doing anything useful if that check failed anyway.)