rawshark: Use the common dissection options
Use the common dissection options processing for rawshark.
This means that the error message for resolving options includes
all the possible resolving options (e.g. instead of omitting VLAN).

This add support for the other options, which generally make sense,
like enabling and disabling protocols and heuristics.

The only option that isn't supported is the Decode As option, only
because '-d' is used by rawshark for its payload link layer type /
dissector selection.