UAT+Qt: Add dissector syntax line edit
Add a syntax line editor for dissector names, that validates and
provides completion. Use it in the UAT configuration.

It has been impossible to use the UAT to set a dissector with
capital letters; change things so that does work.

This doesn't display the description anywhere, but that would be
nice too.

A better fix for #18836.