Restrict allowed characters in dissector names
Dissectors now can have both a name and a long description.
Out of 1722 registered dissector names, 9 used characters not
allowed in field names, with spaces being the worst offender.
Another handful use uppercase characters, which are not allowed
in heuristic dissector names but are allowed in field and protocol
names.

Those characters can create parsing issues (and already do with
UAT table entries for DLT_USER, where it's impossible to select
some of these dissectors).

Restrict the names to the characters allowed in field and protocol
names, renaming those 9 dissectors. Use the new description field
for those, even though it's not displayed everywhere we'd like yet.