XOT: Fix use of tcp_dissect_pdus
The X.25 over TCP dissector is a bit unusual in that it has an
assigned TCP port but also rejects packets that don't match its
heuristic. (The heuristic is very accepting, but that's fine
because it is an assigned port.)

It also uses tcp_dissect_pdus, but sometimes rejects the packets
after tcp_dissect_pdus has asked for one more segment, which
violates the contract for desegmentation. In particular this
causes errors if the initial TCP payload is only 1 byte.

Once it passes the heuristic, just set a conversation dissector and
use tcp_dissect_pdus after that. It's buggy to reject packets with
heuristics when using tcp_dissect_pdus and dissecting across segment
boundaries.

Fix #18919